% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_wellber.R
\name{find.range.wellber}
\alias{find.range.wellber}
\title{Determine range based on choosen country index and plot window size}
\usage{
find.range.wellber(tar_ctry_ind, size)
}
\arguments{
\item{tar_ctry_ind}{Target, or choosen country rank}

\item{size}{Plot window size, either Small, Medium or Large}
}
\value{
Returns a list of sequential country ranking IDs that include the ID for the choosen country
}
\description{
Determine range based on choosen country index and plot window size
}
\examples{
id.range = find.range.wellber(tar_ctry_ind=3, size=11)
}
