% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wellber.R
\name{load_wellber}
\alias{load_wellber}
\title{Load the 2017 OECD Better Life data}
\usage{
load_wellber()
}
\value{
A list containing the 2017 OECD Better Life Index data in \code{\link{tibble}}
form with indicator classes now included
}
\description{
Load the 2017 OECD Better Life data
}
\examples{
data = load_wellber()
}
\seealso{
\code{\link{selector}}, \code{\link{plot}}
}
