% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wellber.R
\name{plot}
\alias{plot}
\title{Plot wellber object}
\usage{
plot(obj, ind_class, ...)
}
\arguments{
\item{obj}{An object of class \code{wellber} output from \code{\link{selector}}}

\item{ind_class}{A flag to indicate whether to plot the indicator or indicator class data}

\item{...}{Other arguments to plot}
}
\value{
Nothing is returned - plot is rendered
}
\description{
Plot wellber object
}
\examples{
data_object = load.wellber()
subset_data = selector(data_object, country = 'Chile',
qual_ind = 'Housing expenditure', samp_type = 'Total', wind_size = 'Small')
plot(subset_data, ind_class = 0)
plot(subset_data, ind_class = 1)
}
\seealso{
\code{\link{load.wellber}}, \code{\link{selector}}
}
