% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp_wellber.R
\name{run}
\alias{run}
\title{Executes app to compared country well-being data}
\usage{
run()
}
\value{
Nothing is returned but the app is run
}
\description{
Executes app to compared country well-being data
}
