% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_wellber.R
\name{selector}
\alias{selector}
\title{Filters well-being data based on country, indicator, sample type and window size}
\usage{
selector(obj, country, qual_ind, samp_type, wind_size)
}
\arguments{
\item{obj}{An object of class \code{wellber} from \code{\link{load.wellber}}}

\item{country}{The selected OECD country to compare against, of which there are 39}

\item{qual_ind}{The selected well-being indicator, of which there are 24}

\item{samp_type}{The sample type to be considered, either Total, Men, Women}

\item{wind_size}{The size of the plot range, either Small, Medium or Large}
}
\value{
Returns a list of class \code{wellber} that includes the filtered indicator dataset, with min and max values, and an indicator class dataset based on the indicator variable selected
}
\description{
Filters well-being data based on country, indicator, sample type and window size
}
\examples{
data_object = load.wellber()
subset_data = selector(data_object, country = 'Chile',
qual_ind = 'Housing expenditure', samp_type = 'Total', wind_size = 'Small')
}
\seealso{
\code{\link{load.wellber}}, \code{\link{plot}}
}
