% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wellber.R
\name{plot}
\alias{plot}
\title{Plot wellber object}
\usage{
plot(obj, num_var, ...)
}
\arguments{
\item{obj}{An object of class \code{wellber} output from \code{\link{selector}}}

\item{num_var}{A flag to indicate whether to plot the indicator or indicator class data}

\item{...}{Other arguments to plot}
}
\value{
Nothing is returned - plot is rendered
}
\description{
Plot wellber object
}
\examples{
data_object = load_wellber()
subset_data = selector(data_object, country = 'Austria',
qual_ind = 'Air pollution', samp_type = 'Total', wind_size = 'small')
plot(subset_data, num_var = 1)
plot(subset_data, num_var = 2)
}
\seealso{
\code{\link{load_wellber}}, \code{\link{selector}}
}
